backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "klugk-backend"
version = "0.1.0"
description = "Django API for the Klug-K kitchen administration platform"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<5.2",
  "djangorestframework>=3.15,<3.16",
  "dj-database-url>=2.2,<3",
  "psycopg[binary]>=3.2,<3.3",
  "PyJWT[crypto]>=2.9,<3",
  "python-dotenv>=1.0,<2",
  "whitenoise>=6.8,<7",
  "gunicorn>=23,<24",
]

[project.optional-dependencies]
dev = [
  "django-stubs>=5.1.0",
  "mypy>=1.12,<2",
  "pytest>=8.3,<9",
  "pytest-django>=4.9,<5",
  "ruff>=0.7,<0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "apps*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
